requirements.txt
absl-py==0.13.0
astor==0.8.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
backports.weakref==1.0rc1
bleach==1.5.0
cached-property==1.5.2
cachetools==4.0.0
certifi==2018.11.29
chardet==3.0.4
click==8.0.3
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.0
defusedxml==0.6.0
dicom==0.9.9.post1
dill==0.3.4
dlx==1.0.4
docplex==2.20.204
entrypoints==0.3
fastdtw==0.3.4
fastjsonschema==2.15.1
flatbuffers==1.12
gast==0.5.3
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.34.1
h5py==2.7.1
html5lib==0.9999999
idna==2.8
imageio==2.5.0
immutables==0.15
importlib-metadata==1.5.0
imutils==0.5.3
inflection==0.5.1
ipykernel==5.1.4
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jax==0.1.3
jaxlib==0.1
jedi==0.16.0
joblib==0.13.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kaggle==1.5.1.1
Keras==2.1.3
Keras-Applications==1.0.8
keras-contrib==2.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Lasagne==0.1
lxml==4.6.3
Markdown==3.1.1
mistune==0.8.4
more-itertools==8.8.0
mpmath==1.2.1
multitasking==0.0.9
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.3
notebook==6.0.3
ntlm-auth==1.5.0
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.3.0
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
Pygments==2.6.1
pylatexenc==2.10
pyrsistent==0.15.7
python-constraint==1.4.0
python-dateutil==2.8.1
python-slugify==2.0.0
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==19.0.0
qiskit==0.27.0
qiskit-aer==0.8.2
qiskit-aqua==0.9.2
qiskit-ibmq-provider==0.14.0
qiskit-ignis==0.6.0
qiskit-terra==0.17.4
qtconsole==4.7.1
QtPy==1.9.0
Quandl==3.6.1
requests==2.21.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
retworkx==0.9.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1629411471490/work
seaborn==0.10.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sympy==1.8
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
tornado==6.0.4
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.7.4.3
Unidecode==1.0.23
urllib3==1.22
wcwidth==0.1.8
websocket-client==1.1.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.12.1
yfinance==0.1.59
zipp==3.1.0
